+++ src/Board.hpp
@@ -12,7 +12,7 @@ class Board {
   Board();
 
   void register_tank();
-  void fire_missle(const float angle, const float x, const float y);
+  void fire_missle(const int angle, const float x, const float y);
 
   void run();
